probe: allow for selection of CMSIS-DAP v1 and v2 protocols
CMSIS-DAP v1 is a pseudo-HID device which allows "driver-less" use on
Windows. DAP v2 is a vendor-specific bulk protocol, so there needs to be
a companion MS OS 2.0 descriptor for the debug interface with
DeviceInterfaceGUID set to the Keil GUID.

Allow for selection between these and the openocd protocol at compile time.

Parts of the v1 implementation are derived from the DapperMime project, so
add credit where due.

Signed-off-by: Jonathan Bell <jonathan@raspberrypi.com>