probe: handle pipelined OUT transfers properly
A consequence of tinyUSB's implementation of vendor interfaces is that
it treats pipe traffic as free-flowing, i.e. no distinction is made if a
short packet is received. CMSIS-DAP should ordinarily encapsulate a set
of commands in a single packet, but if multiple packets are receieved
before the FIFO is drained then all the commands need to be processed at
once.

Ensure there is always enough RX space, and TX space for the generated
response.