Removed glitches on probe startup.
With the previous version cdc_uart characters from target where sent
back to target.
Don't know if that had something to do with cdc flush/clear, but now
this effect is gone.  Price is that the CDC threads are polling for
connection (which is not bad anyway).

Note: propability is just reduced.  I think this has something to do with tinyusb