Add remote_gpio feature and helper app
This provides a framework for tunneling arbitrary GPIO ops through the
USB control pipe.

NOTE: this implementation does not respect probe pin assignments, and
destructive operations can be performed which will require a power cycle
to undo.

It is intended as a proof-of concept for experimentation.

The helper requires libudev-dev and libusb-1.0-dev to be installed on
the host.